"builders refactored; updates #175"
