+++ prototype2/multigrid/mesytec/BuilderMesytec.cpp
@@ -11,7 +11,7 @@ namespace Multigrid {
 
 BuilderMesytec::BuilderMesytec(const SequoiaGeometry &geometry, bool spoof_time,
                                std::string dump_dir)
-    : digital_geometry_(geometry) {
+    : BuilderReadouts(geometry) {
   vmmr16Parser_.spoof_high_time(spoof_time);
   if (!dump_dir.empty()) {
     dumpfile_ = ReadoutFile::create(dump_dir + "mgmesytec_readouts_" + timeString(), 100);
@@ -25,10 +25,10 @@ std::string BuilderMesytec::debug() const {
   ss << "  ========           Mesytec Builder           ========\n";
   ss << "  =====================================================\n";
 
-  ss << "  Geometry mappings:\n";
-  ss << digital_geometry_.debug("  ") << "\n";
   ss << "  Spoof high time (vmmr16):"
      << (vmmr16Parser_.spoof_high_time() ? "YES" : "no") << "\n";
+
+  ss << BuilderReadouts::debug();
 
   return ss.str();
 }
@@ -50,37 +50,7 @@ void BuilderMesytec::parse(Buffer<uint8_t> buffer) {
       dumpfile_->push(vmmr16Parser_.converted_data);
     }
 
-    for (const auto &r : vmmr16Parser_.converted_data) {
-      if (r.external_trigger) {
-        hit_.plane = external_trigger_plane;
-        hit_.coordinate = 0;
-        hit_.weight = 0;
-      } else if (digital_geometry_.isWire(r.bus, r.channel)) {
-        hit_.weight = digital_geometry_.rescale(r.bus, r.channel, r.adc);
-        if (!digital_geometry_.is_valid(r.bus, r.channel, hit_.weight)) {
-          stats_readout_filter_rejects++;
-          continue;
-        }
-        hit_.coordinate = digital_geometry_.wire(r.bus, r.channel);
-        hit_.plane = wire_plane;
-      } else if (digital_geometry_.isGrid(r.bus, r.channel)) {
-        hit_.weight = digital_geometry_.rescale(r.bus, r.channel, r.adc);
-        if (!digital_geometry_.is_valid(r.bus, r.channel, hit_.weight)) {
-          stats_readout_filter_rejects++;
-          continue;
-        }
-        hit_.coordinate = digital_geometry_.grid(r.bus, r.channel);
-        hit_.plane = grid_plane;
-      } else {
-        XTRACE(PROCESS, DEB, "Bad geometry %s", r.debug().c_str());
-        stats_digital_geom_errors++;
-        continue;
-      }
-
-      hit_.time = r.total_time;
-      ConvertedData.push_back(hit_);
-    }
-
+    build(vmmr16Parser_.converted_data);
   }
 }
