+++ prototype2/multigrid/mesytec/CMakeLists.txt
@@ -33,12 +33,44 @@ set(Vmmr16ParserTest_INC
   )
 create_test_executable(Vmmr16ParserTest)
 
-set(TEST_DATA_PATH "${REFDATA}/multigrid/2018_08_30")
+set(TEST_DATA_PATH "${REFDATA}/multigrid/2018_09_03")
 if(EXISTS ${TEST_DATA_PATH})
-  set(BuilderMesytecTest_SRC
-    ../generators/BuilderReadouts.cpp
+  set(MGBuilderReadoutsTest_SRC
+    BuilderReadoutsTest.cpp
+    BuilderReadouts.cpp
     BuilderMesytec.cpp
+    Sis3153Parser.cpp
+    Vmmr16Parser.cpp
+    Readout.cpp
+    ../reduction/EventAnalysis.cpp
+    ../generators/ReaderReadouts.cpp
+    ../Config.cpp
+    )
+  set(MGBuilderReadoutsTest_INC
+    BuilderReadouts.h
+    BuilderMesytec.h
+    Sis3153Parser.h
+    Vmmr16Parser.h
+    Readout.h
+    ../../common/Buffer.h
+    ../geometry/SequoiaGeometry.h
+    ../generators/ReaderReadouts.h
+    ../Config.h
+    )
+  create_test_executable(MGBuilderReadoutsTest)
+  target_compile_definitions(MGBuilderReadoutsTest
+    PRIVATE TEST_DATA_PATH="${TEST_DATA_PATH}/")
+else()
+  message(WARNING "*** No multigrid reference data found. Skipping BuilderReadouts tests.")
+endif()
+
+
+set(TEST_DATA_PATH2 "${REFDATA}/multigrid/2018_08_30")
+if(EXISTS ${TEST_DATA_PATH2})
+  set(BuilderMesytecTest_SRC
     BuilderMesytecTest.cpp
+    BuilderReadouts.cpp
+    BuilderMesytec.cpp
     Vmmr16Parser.cpp
     Sis3153Parser.cpp
     Readout.cpp
@@ -48,12 +80,12 @@ if(EXISTS ${TEST_DATA_PATH})
     )
   set(BuilderMesytecTest_INC
     BuilderMesytec.h
-    ../../common/Buffer.h
     Sis3153Parser.h
     Vmmr16Parser.h
     Readout.h
     TestData.h
     ../geometry/SequoiaGeometry.h
+    ../../common/Buffer.h
     ../../tools/ReaderPcap.h
     ../Config.h
     )
@@ -62,10 +94,10 @@ if(EXISTS ${TEST_DATA_PATH})
     )
   create_test_executable(BuilderMesytecTest)
   target_compile_definitions(BuilderMesytecTest
-    PRIVATE TEST_DATA_PATH="${TEST_DATA_PATH}/")
+    PRIVATE TEST_DATA_PATH="${TEST_DATA_PATH2}/")
   target_compile_definitions(BuilderMesytecTest
     PRIVATE TEST_JSON_PATH="${TEST_JSON_PATH}/")
 else()
-  message(WARNING "*** No multigrid reference data found. Skipping relevant tests.")
+  message(WARNING "*** No multigrid reference data found. Skipping BuilderMesytec tests.")
 endif()
